pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "repository_stats"
version = "0.1.0"
description = "Implements some utilities for github repositories."
readme = "README.md"
authors = [{ name = "Reafel Horowitz", email = "refael1207@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "attrs~=24.2.0",
    "graphviz~=0.20.3",
    "more-itertools~=10.5.0",
    "PyGithub~=2.5.0",
    "pydot~=3.0.2",
    "PyYAML~=6.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.10"
mypy_path = "./src"